requirements.txt
dnspython
sublist3r
beautifulsoup4
requests
selenium
